refactor: someting.utils.ts → someting.util.ts
복수형 폴더 아래에 있는 파일은 단수형으로 표현
