+++ backend/src/common/interceptors/logging.interceptor.ts
@@ -1,7 +1,7 @@
 import { CallHandler, ExecutionContext, Injectable, NestInterceptor } from '@nestjs/common';
 import { Observable, tap } from 'rxjs';
-import { Color, bold, color, italic } from 'src/utils/ansi.utils';
-import { generateUuid } from '../../utils/hash.utils';
+import { Color, bold, color, italic } from 'src/utils/ansi.util';
+import { generateUuid } from '../../utils/hash.util';
 import { WinstonLoggerService } from '../logger/winstonLogger.service';
 
 @Injectable()
