+++ backend/src/s3/s3.service.ts
@@ -3,7 +3,7 @@ import { getSignedUrl } from '@aws-sdk/s3-request-presigner';
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { WinstonLoggerService } from 'src/common/logger/winstonLogger.service';
-import { generateUuid } from 'src/utils/hash.utils';
+import { generateUuid } from 'src/utils/hash.util';
 import { PresignedUrlInfo } from './types/s3.type';
 
 const BUCKET_NAME = 'dangdangwalk';
