+++ backend/src/utils/hash.util.ts
